add new just test projectas
