--- jest-lesson/src/section2/divide.test.ts
@@ -1,7 +1,0 @@
-// toThrow
-
-import { divide, ZeroDivisorError } from './divide';
-
-it("0で割るとエラーになる", () => {
-    expect(() => divide(10, 0)).toThrow()
-});
