--- jest-lesson/src/section2/divide.ts
@@ -1,6 +1,0 @@
-export class ZeroDivisorError extends Error {}
-
-export function divide(dividend: number, divisor: number) {
-  if (divisor === 0) throw new ZeroDivisorError("0で割ることはできません");
-  return dividend / divisor;
-}
