+++ jest-lesson/src/section2/sum.test.ts
@@ -1,5 +1,5 @@
-import { sum } from './sum';
+import { sum } from "./sum";
 
 it("1と2を足すと3になる", () => {
-    expect(sum(1, 2)).toBe(3);
-})
+  expect(sum(1, 2)).toBe(3);
+});
