+++ jest-lesson/src/section2/sum.ts
@@ -1,3 +1,3 @@
-export function sum(a:number, b: number): number {
-    return a + b;
+export function sum(a: number, b: number): number {
+  return a + b;
 }
