--- jest-lesson/src/survival/jest/isZero.ts
@@ -1,3 +1,0 @@
-export function isZero(value: number): boolean {
-    return value === 0;
-}
