--- jest-lesson/src/survival/qiita/app.js
@@ -1,9 +1,0 @@
-const toFizzBuzz = require('./modules/fizzbuzz');
-
-const app = () => {
-    for (let i = 1; i <= 100; i++) {
-        console.log(toFizzBuzz(i));
-    }
-};
-
-module.exports = app;
